Consolidate Strava API interactions and add tests
- Create a few classes to handle the strava API
  in-house, rather than using stravalib.
    - An Oauth client for obtaining/refreshing tokens
    - An API client for obtaining json responses from various endpoints
      using a token from the oauth client
    - A rate limit monitor that can determine the api rate limit status
      from the api responses.
- Add / modify unit tests for stravalib-free API interface.
  - `FlaskTestCase`: add option to persist the db data between tests.
    Facilitates occasional live API testing to ensure it still behaves
    as expected.
- Functional tests are a mess and I've decided to leave them alone
  until the flow is more defined.
- Remove `settings.py` from the tests directory.  Prefer using
  environment variables to override default testing config for now.

+++ tests/unit_tests/base.py
@@ -12,6 +12,8 @@ BASEDIR = os.path.abspath(os.path.dirname(__file__))
 
 
 class FlaskTestCase(unittest.TestCase):
+  clean_db = True
+
   def setUp(self):
     """
     Refs:
@@ -25,10 +27,12 @@ class FlaskTestCase(unittest.TestCase):
     self.app_context = self.app.app_context()
     self.app_context.push()
     self.client = self.app.test_client(use_cookies=True)
-    db.create_all()
+    if self.clean_db:
+      db.create_all()
 
   def tearDown(self):
     db.session.remove()
-    db.drop_all()
+    if self.clean_db:
+      db.drop_all()
     self.app_context.pop()
     self.test_request_context.pop()
